Moved required methods to common module
