+++ backend/common/src/main/java/ai/verta/modeldb/common/handlers/MapSubtypes.java
@@ -1,4 +1,4 @@
-package ai.verta.modeldb.experimentRun.subtypes;
+package ai.verta.modeldb.common.handlers;
 
 import java.util.*;
