+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/FeatureHandler.java
@@ -4,6 +4,7 @@ import ai.verta.modeldb.Feature;
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
 import ai.verta.modeldb.common.futures.InternalFuture;
+import ai.verta.modeldb.common.handlers.MapSubtypes;
 import ai.verta.modeldb.exceptions.InvalidArgumentException;
 import java.util.AbstractMap;
 import java.util.HashSet;
