+++ backend/src/main/java/ai/verta/modeldb/experimentRun/subtypes/TagsHandler.java
@@ -2,32 +2,27 @@ package ai.verta.modeldb.experimentRun.subtypes;
 
 import ai.verta.modeldb.common.exceptions.InternalErrorException;
 import ai.verta.modeldb.common.futures.FutureJdbi;
-import ai.verta.modeldb.common.futures.InternalFuture;
 import ai.verta.modeldb.common.handlers.TagsHandlerBase;
-import ai.verta.modeldb.exceptions.InvalidArgumentException;
-import ai.verta.modeldb.utils.ModelDBUtils;
 import java.util.*;
 import java.util.concurrent.Executor;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 public class TagsHandler extends TagsHandlerBase {
 
   public TagsHandler(Executor executor, FutureJdbi jdbi, String entityName) {
-      super(executor, jdbi, entityName);
+    super(executor, jdbi, entityName);
   }
 
-    @Override
-    protected void setEntityIdReferenceColumn(String entityName) {
-        switch (entityName) {
-            case "ProjectEntity":
-                this.entityIdReferenceColumn = "project_id";
-                break;
-            case "ExperimentRunEntity":
-                this.entityIdReferenceColumn = "experiment_run_id";
-                break;
-            default:
-                throw new InternalErrorException("Invalid entity name: " + entityName);
-        }
+  @Override
+  protected void setEntityIdReferenceColumn(String entityName) {
+    switch (entityName) {
+      case "ProjectEntity":
+        this.entityIdReferenceColumn = "project_id";
+        break;
+      case "ExperimentRunEntity":
+        this.entityIdReferenceColumn = "experiment_run_id";
+        break;
+      default:
+        throw new InternalErrorException("Invalid entity name: " + entityName);
     }
+  }
 }
